pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "Strands Agents Documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click<8.3.0",
    "mike~=2.1.3",
    "mkdocs~=1.6.1",
    "mkdocs-macros-plugin~=1.3.7",
    "mkdocs-material~=9.6.12",
    "mkdocstrings-python~=1.16.10",
    "mkdocs-llmstxt~=0.2.0",
    "strands-agents[all,bidi-all]",
    "pymdown-extensions>=10.16.1",
    "opensearch-py>=2.0.0",
    "retrying>=1.3.3",
    "requests>=2.25.0",
]


[tool.setuptools]
py-modules = []